Stat Changes
Added a new Balance tab (name may not be final) as an Extra tab. You can import a `MeleeWeapons.json` and/or a `RangedWeapons.json` file to check if their stats are different to the stats PegLeg knows about. Importing files can be done with a popup, or by dragging the file into the tab (in which case the file must be correctly named).